@@ -4,37 +4,42 @@
   "private": true,
   "homepage": "https://embed.videosdk.live/rtc-js-prebuilt/0.3.38/",
   "dependencies": {
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
+    "@emoji-mart/react": "^1.1.1",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "@mui/icons-material": "^5.15.15",
+    "@mui/material": "^5.15.15",
+    "@mui/system": "^5.15.15",
     "@videosdk.live/react-sdk": "^0.1.85",
     "@videosdk.live/videosdk-media-processor-web": "^0.0.12",
-    "compressorjs": "^1.1.1",
-    "emoji-mart": "^3.0.1",
-    "fabric": "^4.6.0",
-    "hls.js": "^1.1.5",
+    "compressorjs": "^1.2.1",
+    "emoji-mart": "^5.5.2",
+    "fabric": "^5.3.0",
+    "hls.js": "^1.5.8",
     "i18next": "^22.4.10",
-    "notistack": "^1.0.10",
-    "react": "^17.0.2",
+    "notistack": "^3.0.1",
+    "react": "^18.2.0",
     "react-color": "^2.19.3",
-    "react-device-detect": "^1.17.0",
-    "react-dom": "^17.0.2",
-    "react-emoji-render": "^1.2.4",
+    "react-device-detect": "^2.2.3",
+    "react-dom": "^18.2.0",
+    "react-emoji-render": "^2.0.1",
     "react-i18next": "^12.1.5",
-    "react-lines-ellipsis": "^0.15.0",
+    "react-lines-ellipsis": "^0.15.4",
     "react-linkify": "0.2.2",
     "react-lottie": "^1.2.3",
-    "react-motion": "^0.5.2",
-    "react-player": "^2.9.0",
-    "react-responsive": "^9.0.0-beta.4",
-    "react-scripts": "4.0.3",
-    "react-use-clipboard": "^1.0.7",
-    "react-virtualized": "^9.22.3",
+    "react-player": "^2.16.0",
+    "react-responsive": "^9.0.2",
+    "react-scripts": "^5.0.1",
+    "react-spring": "^9.7.3",
+    "react-use-clipboard": "^1.0.9",
+    "react-virtualized": "^9.22.5",
     "react-visibility-sensor": "^5.1.1",
     "uuid": "^8.3.2",
-    "web-vitals": "^1.0.1"
+    "web-vitals": "^1.1.2"
+  },
+  "overrides": {
+    "nth-check": "2.1.1",
+    "postcss": "^8.4.31"
   },
   "scripts": {
     "start": "react-scripts start",
