@@ -10,8 +10,8 @@ import {
 } from "./MeetingAppContextDef";
 import JoinMeeting from "./components/JoinScreen";
 import ClickAnywhereToContinue from "./components/ClickAnywhereToContinue";
-import { Box, CircularProgress, useMediaQuery } from "@material-ui/core";
-import { useTheme } from "@material-ui/core/styles";
+import { Box, CircularProgress, useMediaQuery } from "@mui/material";
+import { useTheme } from "@mui/system";
 import MeetingLeftScreen from "./components/MeetingLeftScreen";
 import ConfirmBox from "./components/ConfirmBox";
 import {
@@ -23,7 +23,7 @@ import {
 import useIsSMDesktop from "./utils/useIsSMDesktop";
 import useIsLGDesktop from "./utils/useIsLGDesktop";
 import useIsTab from "./utils/useIsTab";
-import { version as prebuiltSDKVersion } from "../package.json";
+import packageInfo from '../package.json';
 import { meetingModes } from "./CONSTS";
 import animationData from "./animations/meeting-left.json";
 import lightThemeAnimationData from "./animations/meeting_left_white.json";
@@ -31,6 +31,7 @@ import i18n from "i18next";
 import { initReactI18next } from "react-i18next";
 
 const App = () => {
+  const prebuiltSDKVersion = packageInfo.version;
   const [meetingIdValidation, setMeetingIdValidation] = useState({
     isLoading: true,
     meetingId: null,
