@@ -1,5 +1,5 @@
-import { Box, useTheme } from "@material-ui/core";
-import { useEffect, useMemo, useState } from "react";
+import { Box, useTheme } from "@mui/material";
+import { useEffect, useMemo } from "react";
 import Lottie from "react-lottie";
 import useResponsiveSize from "../../utils/useResponsiveSize";
 import animationData from "../../../src/animations/wait_for_HLS_animation.json";
